clearcode-toolkit/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.7            # via django
attrs==19.3.0             # via clearcode-toolkit (setup.py)
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
click==7.1.1              # via clearcode-toolkit (setup.py)
django==3.0.7             # via clearcode-toolkit (setup.py), djangorestframework
djangorestframework==3.11.0  # via clearcode-toolkit (setup.py)
idna==2.9                 # via requests
psycopg2==2.8.5           # via clearcode-toolkit (setup.py)
pytz==2019.3              # via django
requests==2.23.0          # via clearcode-toolkit (setup.py)
sqlparse==0.3.1           # via django
urllib3==1.25.9           # via requests
